tests: hoist the company seed/clock helpers into test_company_cleanup.h
newest_company_stamp(), seed_open_company() and CompanyClockRestore were
written as file-local statics in test_hire_team.cpp for the two hire
flows. Six more og_test_menu_ui flows (BackToMainmenu, OptionsMenu,
NetworkingMenu, Difficulty, Help, LevelProgress) make the same unchecked
assumption the hire flows made — that CONTINUE opens the company the
test just wrote — and need the same three lines, so the helpers move to
the shared company-hygiene header beside ScopedCompanyFileCleanup rather
than being copied six times (the "no rule twins" principle, PR #245).

Pure move: the three comments travel verbatim, `static` becomes
`inline`, and test_hire_team.cpp loses the definitions (it already
includes the header, so keeping them would be a duplicate definition).
No behaviour change; no src/ change.
